ruuvitag_latest/requirements.txt
google-cloud-bigquery==2.5.0
